refactor: refatora autocomplete
